+++ packages/arcodesign/components/sticky/__ast__/index.ast.json
@@ -314,13 +314,13 @@
             "defaultValue": {
                 "value": "() => window"
             },
-            "description": "指定滚动容器，如果指定该值则relative属性始终认定为false；如果返回string则使用querySelector选取容器\n@en Specifies the scrolling container. If this value is specified, the relative property is always regarded as false; if a string is input, use querySelector to select the container",
+            "description": "指定滚动容器；如果返回string则使用querySelector选取容器\n@en Specifies the scrolling container. If this value is specified, the relative property is always regarded as false; if a string is input, use querySelector to select the container",
             "name": "getScrollContainer",
             "tags": {
                 "en": "Specifies the scrolling container. If this value is specified, the relative property is always regarded as false; if a string is input, use querySelector to select the container",
                 "default": "() => window"
             },
-            "descWithTags": "指定滚动容器，如果指定该值则relative属性始终认定为false；如果返回string则使用querySelector选取容器",
+            "descWithTags": "指定滚动容器；如果返回string则使用querySelector选取容器",
             "parent": {
                 "fileName": "arcom-github/packages/arcodesign/components/sticky/index.tsx",
                 "name": "StickyProps"
